requirements.txt
acres==0.3.0
aiosqlite==0.21.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
billiard==4.2.1
Brotli==1.1.0
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
configobj==5.0.9
configparser==7.2.0
cssselect2==0.8.0
distro==1.9.0
etelemetry==0.3.1
fastapi==0.115.12
filelock==3.18.0
flower==2.0.1
fonttools==4.58.0
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
humanize==4.12.3
idna==3.10
isodate==0.6.1
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.5.3
langchain==0.3.25
langchain-core==0.3.60
langchain-openai==0.3.17
langchain-text-splitters==0.3.8
langsmith==0.3.42
looseversion==1.3.0
lxml==5.4.0
MarkupSafe==3.0.2
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.6
openai==1.79.0
orjson==3.10.18
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow==11.2.1
prometheus_client==0.22.0
prompt_toolkit==3.0.51
prov==2.0.1
puremagic==1.29
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydot==4.0.0
pydyf==0.11.0
PyMuPDF==1.25.5
pyparsing==3.2.3
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==6.3.2
redis==6.1.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
scipy==1.15.3
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tenacity==9.1.2
tiktoken==0.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tornado==6.5.1
tqdm==4.67.1
traits==7.0.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
vine==5.1.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
zstandard==0.23.0
